+++ src/delfino/commands/verify_all.py
@@ -1,16 +1,34 @@
+from typing import Dict, cast
+
 import click
 
-from delfino.click_utils.command import command_names
+from delfino.click_utils.command import get_root_command
 from delfino.commands.format import run_format
 from delfino.commands.lint import lint
 from delfino.commands.test import test_all
 from delfino.commands.typecheck import typecheck
+from delfino.contexts import AppContext, pass_app_context
 
 _COMMANDS = [run_format, lint, typecheck, test_all]
 
 
-@click.command(help=f"Runs all checks.\n\nAlias for the {command_names(_COMMANDS)} commands.")
+@click.command(help="Runs all verification commands. Configured by the ``verify_commands`` setting.")
 @click.pass_context
-def verify_all(click_context: click.Context):
-    for command in _COMMANDS:
+@pass_app_context
+def verify_all(click_context: click.Context, app_context: AppContext):
+    delfino = app_context.pyproject_toml.tool.delfino
+
+    root = get_root_command(click_context)
+    commands: Dict[str, click.Command] = {
+        command: cast(click.Command, root.get_command(click_context, command))
+        for command in root.list_commands(click_context)
+    }
+
+    target_commands = [
+        commands[target_name]
+        for target_name in delfino.verify_commands
+        if target_name in commands and target_name not in delfino.disable_commands
+    ]
+
+    for command in target_commands:
         click_context.forward(command)