+++ src/delfino/models/pyproject_toml.py
@@ -1,5 +1,5 @@
 from pathlib import Path
-from typing import Any, Dict, List, Optional, Set
+from typing import Any, Dict, List, Optional, Set, Tuple
 
 from pydantic import BaseModel, Extra, Field
 
@@ -15,10 +15,9 @@ class Delfino(BaseModel):
     reports_directory: Path = Path("reports")
     test_types: List[str] = ["unit", "integration"]
     disable_commands: Set[str] = Field(default_factory=set)
+    verify_commands: Tuple[str, ...] = ("format", "lint", "typecheck", "test-all")
     disable_pre_commit: bool = False
-
     dockerhub: Optional[Dockerhub] = None
-
     plugins: Dict[str, Any] = Field(default_factory=dict, description="Any additional config given by plugins.")
 
     class Config: